refactor: update video with text section
